microservices-jaeger/auth_service/requirements.txt
flask
gunicorn
opentelemetry-instrumentation-flask
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http